+++ packages/prettier-plugin-apex/tests/comments_trailing_with_final_newline/jsfmt.spec.ts
@@ -1,0 +1,6 @@
+import { fileURLToPath } from "url";
+
+runSpec(fileURLToPath(new URL(".", import.meta.url)), ["apex"], {
+  tabWidth: 2,
+  apexInsertFinalNewline: true,
+});